+++ main.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 import codecs
+import datetime
 import sys
 import time
 from tabulate import tabulate
@@ -164,6 +165,99 @@ def match_algorithm(imdb, current_fa_movie):
         if fa_id in hard_coded_matches:
             imdbID = imdb.get_from_code(hard_coded_matches[fa_id])
     return imdbID
+
+
+def progress_format(total, current):
+    return '{}/{}={}%'.format(current, total, 100 * current / total)
+    pass
+
+
+class Eta:
+    def __get_now(self):
+        return datetime.datetime.now()
+
+    def __init__(self, total):
+        self._total = total
+        self._init_time = self.__get_now()
+        pass
+
+    def set_current(self, current):
+        self._current = current
+
+    def format(self):
+        elapsed = (self.__get_now() - self._init_time)
+        tot_time = elapsed * self._total / self._current
+        pend_time = tot_time - elapsed
+        return '{} {}/{}'.format(progress_format(current=self._current, total=self._total), pend_time, tot_time)
+
+
+class Counters:
+    def __init__(self, total):
+        self.a1hits = 0
+        self.a2hits = 0
+        self.count = 0
+        self.eta = Eta(total)
+
+    def format(self):
+        self.eta.set_current(self.count)
+        return 'progress={} alg1hits={} alg2hits={}'.format(self.eta.format(),
+                                                            progress_format(current=self.a1hits, total=self.count),
+                                                            progress_format(current=self.a2hits, total=self.count))
+
+
+def algorithm_strict(imdb, current_fa_movie, counters):
+    alg1 = imdb.match_algorithm(current_fa_movie.movieTitle, current_fa_movie.movieYear)
+    hcmc = hard_coded_matches.get(current_fa_movie.movieID)
+    alg2 = imdb.match_algorithm_new(current_fa_movie.movieTitle, current_fa_movie.movieYear)
+
+    found1 = alg1.could_match()
+    found2 = alg2 is not None and alg2.could_match()
+    strict = True
+    result = None
+    if hcmc is None:
+
+        if found1:
+            counters.a1hits += 1
+            result = alg1
+        else:
+            pass
+
+        if found2:
+            counters.a2hits += 1
+            result = alg2
+        else:
+            pass
+
+        if found1 and found2 and not alg1.get_code() == alg2.get_code():
+            print(
+                'ERROR: algorithms 1 and 2 do not match, please, add {} to hard coded match table to increase matching efficiency.'.format(
+                    current_fa_movie.movieTitle))
+            if strict:
+                assert False
+    else:
+
+        if found1:
+            if alg1.get_code() == hcmc:
+                counters.a1hits += 1
+                result = alg1
+            else:
+                pass
+
+        if found2:
+            if alg2.get_code() == hcmc:
+                counters.a2hits += 1
+                result = alg2
+            else:
+                if result is None:
+                    pass
+                    if strict:
+                        assert False
+
+        else:
+            pass
+
+    pass
+    return result
 
 
 def movie_match(current_fa_movie, imdb, match_results, imdbNotFound):
